+++ deltafi-ui/src/pages/VersionsPage.vue
@@ -1,26 +1,20 @@
 <template>
   <div>
     <PageHeader heading="Versions" />
-    <CollapsiblePanel header="Image Versions" class="table-panel">
-      <DataTable responsive-layout="scroll" :value="versions" striped-rows class="p-datatable-sm p-datatable-gridlines" :loading="showLoading">
-        <template #empty>No version information available.</template>
-        <template #loading>Loading version information. Please wait.</template>
-        <Column field="app" header="App" :sortable="true" />
-        <Column field="container" header="Container" :sortable="true" />
-        <Column field="image.name" header="Image" :sortable="true" />
-        <Column field="image.tag" header="Tag" :sortable="true" />
-      </DataTable>
-    </CollapsiblePanel>
+    <ProgressBar v-if="showLoading" mode="indeterminate" style="height: 0.5em" />
+    <div v-for="group in groups" v-else :key="group">
+      <VersionsPanel v-if="verionsByGroup[group]" class="mb-3" :header="group" :versions="verionsByGroup[group]"></VersionsPanel>
+    </div>
   </div>
 </template>
 
 <script setup>
-import Column from "primevue/column";
-import DataTable from "primevue/datatable";
-import CollapsiblePanel from "@/components/CollapsiblePanel.vue";
+import ProgressBar from "primevue/progressbar";
+import VersionsPanel from "@/components/VersionsPanel.vue";
 import PageHeader from "@/components/PageHeader.vue";
 import useVersions from "@/composables/useVersions";
 import { onMounted, computed, onUnmounted, inject } from "vue";
+import _ from 'lodash';
 
 const refreshInterval = 5000; // 5 seconds
 const { data: versions, loaded, loading, fetch: fetchVersions } = useVersions();
@@ -29,6 +23,20 @@ const isIdle = inject("isIdle");
 let autoRefresh;
 
 const showLoading = computed(() => !loaded.value && loading.value);
+
+const groups = ['Core', 'Plugins', 'Other'];
+const verionsByGroup = computed(() => {
+  return _.groupBy(versions.value, (v) => {
+    switch (v.group) {
+      case 'deltafi-core':
+        return 'Core'
+      case 'deltafi-plugins':
+        return 'Plugins'
+      default:
+        return 'Other'
+    };
+  });
+})
 
 onMounted(() => {
   fetchVersions();